Move main.py to top-level folder to fix import issues
This is a temporary fix. Because of the relative imports, it has to be run as a package, which is not the case for the planit source folder when you execute the main.py file directly.

Perhaps I can move it back and then simply run it as a module via python -m ...