Switch to the fancy ListBox impl. for the search engine preview
To make the results window of the search engine look nicer and more customizable, I've moved away from the standard ListBox widget and moved to the custom FancyListBox widget.

It uses the SegmentedControl class to manage the selection of the buttons, similarly to how the toolbar manages the selection of the tools.

Implements most of the UI of #16